"cleanup and restructuring; updates #175"
