+++ prototype2/multigrid/CMakeLists.txt
@@ -10,33 +10,33 @@ file(COPY ${CMAKE_CURRENT_SOURCE_DIR}/calib_data
 # Two-thread implementation of Multi Grid demonstrator detector
 # for Mesytec Readout
 set(mgmesytec_common_src
-  MGMesytecBase.cpp
   ../common/Hists.cpp
   ../common/HistSerializer.cpp
   ../common/ReadoutSerializer.cpp
-  MgConfig.cpp
+  parsers/Vmmr16Parser.cpp
+  parsers/Sis3153Parser.cpp
+  parsers/Hit.cpp
   mgmesytec/EfuMaximum.cpp
   mgmesytec/EfuCenterMass.cpp
   mgmesytec/EfuPrioritized.cpp
-  mgmesytec/Vmmr16Parser.cpp
-  mgmesytec/Sis3153Parser.cpp
-  mgmesytec/Hit.cpp
+  MgConfig.cpp
+  MGMesytecBase.cpp
   )
 set(mgmesytec_common_inc
-  MGMesytecBase.h
   ../common/Hists.h
   ../common/HistSerializer.h
   ../common/ReadoutSerializer.h
-  MgConfig.h
+  parsers/Vmmr16Parser.h
+  parsers/Sis3153Parser.h
+  parsers/Hit.h
   mgmesytec/Efu.h
   mgmesytec/EfuMaximum.h
   mgmesytec/EfuCenterMass.h
   mgmesytec/EfuPrioritized.h
-  mgmesytec/Vmmr16Parser.h
-  mgmesytec/Sis3153Parser.h
-  mgmesytec/Hit.h
   mgmesytec/SequoiaGeometry.h
   mgmesytec/BusGeometry.h
+  MGMesytecBase.h
+  MgConfig.h
   )
 
 set(mgmesytec_INC
@@ -53,6 +53,20 @@ create_module(mgmesytec)
 # Tests
 #=============================================================================
 
+add_subdirectory(parsers)
+add_subdirectory(mgmesytec)
+
+set(MgConfigTest_SRC
+  MgConfig.cpp
+  MgConfigTest.cpp
+  ../../prototype2/common/DataSave.cpp
+  )
+set(MgConfigTest_INC
+  MgConfig.h
+  ../../prototype2/common/DataSave.h
+  )
+create_test_executable(MgConfigTest)
+
 set(MGMesytecBaseTest_INC
   ${mgmesytec_common_inc}
   ../../libs/include/Socket.h
@@ -67,118 +81,3 @@ set(MGMesytecBaseTest_SRC
   ../adc_readout/test/TestUDPServer.cpp
 )
 create_test_executable(MGMesytecBaseTest)
-
-set(MgConfigTest_SRC
-  MgConfig.cpp
-  MgConfigTest.cpp
-  ../../prototype2/common/DataSave.cpp
-)
-set(MgConfigTest_INC
-  MgConfig.h
-  ../../prototype2/common/DataSave.h
-)
-create_test_executable(MgConfigTest)
-
-set(MGHitTest_SRC
-  mgmesytec/HitTest.cpp
-  mgmesytec/Hit.cpp
-  )
-set(MGHitTest_INC
-  mgmesytec/Hit.h
-  )
-create_test_executable(MGHitTest SKIP_MEMGRIND)
-
-set(Sis3153ParserTest_SRC
-  mgmesytec/Sis3153ParserTest.cpp
-  mgmesytec/Sis3153Parser.cpp
-  )
-set(Sis3153ParserTest_INC
-  mgmesytec/Sis3153Parser.h
-  mgmesytec/TestData.h
-  ../common/Buffer.h
-  )
-create_test_executable(Sis3153ParserTest)
-
-set(Vmmr16ParserTest_SRC
-  mgmesytec/Vmmr16ParserTest.cpp
-  mgmesytec/Vmmr16Parser.cpp
-  mgmesytec/Sis3153Parser.cpp
-  mgmesytec/Hit.cpp
-  )
-set(Vmmr16ParserTest_INC
-  ../common/Buffer.h
-  mgmesytec/Sis3153Parser.h
-  mgmesytec/Vmmr16Parser.h
-  mgmesytec/Hit.h
-  mgmesytec/TestData.h
-  )
-create_test_executable(Vmmr16ParserTest)
-
-set(MG24GeometryTest_SRC
-  mgmesytec/MG24GeometryTest.cpp)
-set(MG24GeometryTest_INC
-  mgmesytec/MG24Geometry.h)
-create_test_executable(MG24GeometryTest)
-
-set(FilterTest_SRC
-  mgmesytec/FilterTest.cpp)
-set(FilterTest_INC
-  mgmesytec/Filter.h)
-create_test_executable(FilterTest)
-
-set(BusGeometryTest_SRC
-  mgmesytec/BusGeometryTest.cpp)
-set(BusGeometryTest_INC
-  mgmesytec/BusGeometry.h)
-create_test_executable(BusGeometryTest)
-
-set(SequoiaGeometryTest_SRC
-  mgmesytec/SequoiaGeometryTest.cpp)
-set(SequoiaGeometryTest_INC
-  mgmesytec/SequoiaGeometry.h)
-create_test_executable(SequoiaGeometryTest)
-
-set(EfuMaximumTest_SRC
-  mgmesytec/EfuMaximumTest.cpp
-  mgmesytec/EfuMaximum.cpp
-  mgmesytec/Hit.cpp
-  ../common/Hists.cpp
-  ../common/ReadoutSerializer.cpp)
-set(EfuMaximumTest_INC
-  mgmesytec/EfuMaximum.h
-  mgmesytec/Efu.h
-  mgmesytec/SequoiaGeometry.h
-  mgmesytec/Hit.h
-  ../common/Hists.h
-  ../common/ReadoutSerializer.h)
-create_test_executable(EfuMaximumTest)
-
-set(EfuCenterMassTest_SRC
-  mgmesytec/EfuCenterMassTest.cpp
-  mgmesytec/EfuCenterMass.cpp
-  mgmesytec/Hit.cpp
-  ../common/Hists.cpp
-  ../common/ReadoutSerializer.cpp)
-set(EfuCenterMassTest_INC
-  mgmesytec/EfuCenterMass.h
-  mgmesytec/Efu.h
-  mgmesytec/SequoiaGeometry.h
-  mgmesytec/Hit.h
-  ../common/Hists.h
-  ../common/ReadoutSerializer.h)
-create_test_executable(EfuCenterMassTest)
-
-set(EfuPrioritizedTest_SRC
-  mgmesytec/EfuPrioritizedTest.cpp
-  mgmesytec/EfuPrioritized.cpp
-  mgmesytec/Hit.cpp
-  ../common/Hists.cpp
-  ../common/ReadoutSerializer.cpp)
-set(EfuPrioritizedTest_INC
-  mgmesytec/EfuCenterMass.h
-  mgmesytec/Efu.h
-  mgmesytec/SequoiaGeometry.h
-  mgmesytec/Hit.h
-  ../common/Hists.h
-  ../common/ReadoutSerializer.h)
-create_test_executable(EfuPrioritizedTest)
