+++ prototype2/multigrid/parsers/Hit.cpp
@@ -1,6 +1,6 @@
 /** Copyright (C) 2016, 2017 European Spallation Source ERIC */
 
-#include <multigrid/mgmesytec/Hit.h>
+#include <multigrid/parsers/Hit.h>
 #include <sstream>
 
 namespace Multigrid {
