+++ prototype2/multigrid/parsers/Sis3153ParserTest.cpp
@@ -1,7 +1,7 @@
 /** Copyright (C) 2016, 2017 European Spallation Source ERIC */
 
-#include <multigrid/mgmesytec/Sis3153Parser.h>
-#include <multigrid/mgmesytec/TestData.h>
+#include <multigrid/parsers/Sis3153Parser.h>
+#include <multigrid/parsers/TestData.h>
 #include <test/TestBase.h>
 
 using namespace Multigrid;
