+++ prototype2/multigrid/parsers/Vmmr16ParserTest.cpp
@@ -1,8 +1,8 @@
 /** Copyright (C) 2016, 2017 European Spallation Source ERIC */
 
-#include <multigrid/mgmesytec/Sis3153Parser.h>
-#include <multigrid/mgmesytec/Vmmr16Parser.h>
-#include <multigrid/mgmesytec/TestData.h>
+#include <multigrid/parsers/Sis3153Parser.h>
+#include <multigrid/parsers/Vmmr16Parser.h>
+#include <multigrid/parsers/TestData.h>
 #include <test/TestBase.h>
 
 using namespace Multigrid;
